fix: remove scrolling, and move cursor container to cover the entire page
