+++ packages/nextjs/components/cursor/LiveCursorContainer.tsx
@@ -113,7 +113,7 @@ export const LiveCursorContainer = ({ children }: any) => {
       window.removeEventListener("keyup", onKeyUp);
       window.removeEventListener("keydown", onKeyDown);
     };
-  }, [updateMyPresence]);
+  }, [updateMyPresence, cursor]);
 
   useEventListener(eventData => {
     const event = eventData.event as ReactionEvent;
@@ -129,136 +129,134 @@ export const LiveCursorContainer = ({ children }: any) => {
   });
 
   return (
-    <>
-      <div
-        className="relative h-screen w-full flex items-center justify-center overflow-hidden touch-none"
-        style={{
-          cursor: state.mode === CursorMode.Chat ? "none" : "url(cursor.svg) 0 0, auto",
-        }}
-        onPointerMove={event => {
-          event.preventDefault();
-          if (cursor == null || state.mode !== CursorMode.ReactionSelector) {
-            updateMyPresence({
-              cursor: {
-                x: Math.round(event.clientX),
-                y: Math.round(event.clientY),
-              },
-            });
-          }
-        }}
-        onPointerLeave={() => {
-          setState({
-            mode: CursorMode.Hidden,
-          });
-          updateMyPresence({
-            cursor: null,
-          });
-        }}
-        onPointerDown={event => {
+    <div
+      className=" overflow-hidden touch-none"
+      style={{
+        cursor: state.mode === CursorMode.Chat ? "none" : "url(cursor.svg) 0 0, auto",
+      }}
+      onPointerMove={event => {
+        event.preventDefault();
+        if (cursor == null || state.mode !== CursorMode.ReactionSelector) {
           updateMyPresence({
             cursor: {
               x: Math.round(event.clientX),
               y: Math.round(event.clientY),
             },
           });
-          setState(state => (state.mode === CursorMode.Reaction ? { ...state, isPressed: true } : state));
-        }}
-        onPointerUp={() => {
-          setState(state => (state.mode === CursorMode.Reaction ? { ...state, isPressed: false } : state));
-        }}
-      >
-        {reactions.map(reaction => {
-          return (
-            <FlyingReaction
-              key={reaction.timestamp.toString()}
-              x={reaction.point.x}
-              y={reaction.point.y}
-              timestamp={reaction.timestamp}
-              value={reaction.value}
-            />
-          );
-        })}
-        {cursor && (
-          <div
-            className="absolute top-0 left-0"
-            style={{
-              transform: `translateX(${cursor.x}px) translateY(${cursor.y}px)`,
-            }}
-          >
-            {state.mode === CursorMode.Chat && (
-              <>
-                <Image width="24" height="36" src="cursor.svg" alt="cursor" />
+        }
+      }}
+      onPointerLeave={() => {
+        setState({
+          mode: CursorMode.Hidden,
+        });
+        updateMyPresence({
+          cursor: null,
+        });
+      }}
+      onPointerDown={event => {
+        updateMyPresence({
+          cursor: {
+            x: Math.round(event.clientX),
+            y: Math.round(event.clientY),
+          },
+        });
+        setState(state => (state.mode === CursorMode.Reaction ? { ...state, isPressed: true } : state));
+      }}
+      onPointerUp={() => {
+        setState(state => (state.mode === CursorMode.Reaction ? { ...state, isPressed: false } : state));
+      }}
+    >
+      {reactions.map(reaction => {
+        return (
+          <FlyingReaction
+            key={reaction.timestamp.toString()}
+            x={reaction.point.x}
+            y={reaction.point.y}
+            timestamp={reaction.timestamp}
+            value={reaction.value}
+          />
+        );
+      })}
+      {cursor && (
+        <div
+          className="absolute top-0 left-0 z-40"
+          style={{
+            transform: `translateX(${cursor.x}px) translateY(${cursor.y}px)`,
+          }}
+        >
+          {state.mode === CursorMode.Chat && (
+            <>
+              <Image width="24" height="36" src="cursor.svg" alt="cursor" />
 
-                <div
-                  className="absolute top-5 left-2 px-4 py-2 bg-blue-500 text-white leading-relaxed text-sm"
-                  onKeyUp={e => e.stopPropagation()}
-                  style={{
-                    borderRadius: 20,
+              <div
+                className="absolute top-5 left-2 px-4 py-2 bg-blue-500 text-white leading-relaxed text-sm"
+                onKeyUp={e => e.stopPropagation()}
+                style={{
+                  borderRadius: 20,
+                }}
+              >
+                {state.previousMessage && <div>{state.previousMessage}</div>}
+                <input
+                  className="bg-transparent border-none	outline-none text-white placeholder-blue-300 w-60"
+                  autoFocus={true}
+                  onChange={e => {
+                    updateMyPresence({ message: e.target.value });
+                    setState({
+                      mode: CursorMode.Chat,
+                      previousMessage: null,
+                      message: e.target.value,
+                    });
                   }}
-                >
-                  {state.previousMessage && <div>{state.previousMessage}</div>}
-                  <input
-                    className="bg-transparent border-none	outline-none text-white placeholder-blue-300 w-60"
-                    autoFocus={true}
-                    onChange={e => {
-                      updateMyPresence({ message: e.target.value });
+                  onKeyDown={e => {
+                    if (e.key === "Enter") {
                       setState({
                         mode: CursorMode.Chat,
-                        previousMessage: null,
-                        message: e.target.value,
+                        previousMessage: state.message,
+                        message: "",
                       });
-                    }}
-                    onKeyDown={e => {
-                      if (e.key === "Enter") {
-                        setState({
-                          mode: CursorMode.Chat,
-                          previousMessage: state.message,
-                          message: "",
-                        });
-                      } else if (e.key === "Escape") {
-                        setState({
-                          mode: CursorMode.Hidden,
-                        });
-                      }
-                    }}
-                    placeholder={state.previousMessage ? "" : "Say something…"}
-                    value={state.message}
-                    maxLength={50}
-                  />
-                </div>
-              </>
-            )}
-            {state.mode === CursorMode.ReactionSelector && (
-              <ReactionSelector
-                setReaction={reaction => {
-                  setReaction(reaction);
-                }}
-              />
-            )}
-            {state.mode === CursorMode.Reaction && (
-              <div className="absolute top-3.5 left-1 pointer-events-none select-none">{state.reaction}</div>
-            )}
-          </div>
-        )}
-
-        {others.map(({ connectionId, presence }) => {
-          if (presence == null || !presence.cursor) {
-            return null;
-          }
-          return (
-            <>
-              <Cursor
-                key={connectionId}
-                color={COLORS[connectionId % COLORS.length]}
-                x={presence.cursor.x}
-                y={presence.cursor.y}
-                message={presence.message}
-              />
+                    } else if (e.key === "Escape") {
+                      setState({
+                        mode: CursorMode.Hidden,
+                      });
+                    }
+                  }}
+                  placeholder={state.previousMessage ? "" : "Say something…"}
+                  value={state.message}
+                  maxLength={50}
+                />
+              </div>
             </>
-          );
-        })}
-        {children}
-      </div>
-    </>
+          )}
+          {state.mode === CursorMode.ReactionSelector && (
+            <ReactionSelector
+              setReaction={reaction => {
+                setReaction(reaction);
+              }}
+            />
+          )}
+          {state.mode === CursorMode.Reaction && (
+            <div className="absolute top-3.5 left-1 pointer-events-none select-none">{state.reaction}</div>
+          )}
+        </div>
+      )}
+
+      {others.map(({ connectionId, presence }) => {
+        if (presence == null || !presence.cursor) {
+          return null;
+        }
+        return (
+          <>
+            <Cursor
+              key={connectionId}
+              color={COLORS[connectionId % COLORS.length]}
+              x={presence.cursor.x}
+              y={presence.cursor.y}
+              message={presence.message}
+            />
+          </>
+        );
+      })}
+      {children}
+    </div>
   );
 };
